Refactor imports and remove unused constants; update paths to use aliases
