+++ src/app.tsx
@@ -8,8 +8,8 @@ import './index.css';
 import { MantineProvider } from '@mantine/core';
 import { Notifications } from '@mantine/notifications';
 import { AppProvider } from '@features/app-context';
-import { theme } from './theme/theme';
 
+import { theme } from '@theme/theme';
 import { Router } from './router/router';
 
 export default function App() {
