+++ src/components/settings-modal/settings-modal.tsx
@@ -2,7 +2,7 @@ import { useAppNotifications } from '@components/app-notifications';
 import { useAppContext } from '@features/app-context';
 import { Text, Modal, Stack, Button, Group } from '@mantine/core';
 import { openDB } from 'idb';
-import { FILE_HANDLE_DB_NAME, FILE_HANDLE_STORE_NAME } from '../../consts';
+import { FILE_HANDLE_DB_NAME, FILE_HANDLE_STORE_NAME } from '@consts/idb';
 import { clearFileSystem } from './utils';
 
 interface SettingsModalProps {
