+++ src/components/settings-modal/utils.ts
@@ -1,10 +1,10 @@
-import { openDB } from 'idb';
 import {
   FILE_HANDLE_DB_NAME,
   FILE_HANDLE_STORE_NAME,
   TABS_DB_NAME,
   TABS_STORE_NAME,
-} from '../../consts';
+} from '@consts/idb';
+import { openDB } from 'idb';
 
 export const clearFileSystem = async () => {
   const root = await navigator.storage.getDirectory();
