+++ src/features/app-context/utils.ts
@@ -1,11 +1,11 @@
 /* eslint-disable no-continue */
-import { DataBaseModel } from 'models';
 
 import { EditorState } from '@uiw/react-codemirror';
 import { PostgreSQL, sql } from '@codemirror/lang-sql';
 import { syntaxTree } from '@codemirror/language';
 import { tableFromIPC } from 'apache-arrow';
 import { createName } from '@utils/helpers';
+import { DataBaseModel } from '@models/common';
 import { splitSqlQuery } from '../../utils/editor/statement-parser';
 import { DBRunQueryProps, DBWorkerAPIType, RunQueryResponse, SessionFiles } from './models';
