--- src/models/table.ts
@@ -1,1 +1,0 @@
-export type Limit = 100 | 1000 | 10000;
