+++ tsconfig.json
@@ -23,7 +23,11 @@
       "@utils/*": ["utils/*"],
       "@features/*": ["features/*"],
       "@store/*": ["store/*"],
-      "@hooks/*": ["hooks/*"]
+      "@hooks/*": ["hooks/*"],
+      "@theme/*": ["theme/*"],
+      "@router/*": ["router/*"],
+      "@consts/*": ["consts/*"],
+      "@models/*": ["models/*"]
     }
   },
   "include": ["src"]
